Add compression tests for data handling and default configuration
