+++ src/tests/middleware/mod.rs
@@ -1,830 +1,6 @@
+pub mod body;
 pub mod compression;
-
-#[cfg(test)]
-mod tests {
-
-    use crate::middlewares::body_limit::body_limit;
-    use crate::middlewares::compression::{
-        CompressionConfig, compress_data, should_compress_content_type,
-    };
-    use crate::middlewares::file_upload::FileUploadConfiguration;
-    use crate::middlewares::rate_limiter::{RateLimiterConfig, rate_limiter};
-    use crate::req::body::{RequestBody, RequestBodyContent};
-    use crate::req::request_headers::RequestHeaders;
-    use crate::res::response_status::StatusCode;
-    use crate::{
-        context::{HttpRequest, HttpResponse},
-        middlewares::{
-            cors::{CorsConfig, cors},
-            file_upload::file_upload,
-            logger::{LoggerConfig, logger},
-        },
-        types::HttpMethods,
-    };
-    use std::collections::HashMap;
-    use std::fs;
-    use std::net::IpAddr;
-    use std::str::FromStr;
-    use std::time::Duration;
-    use tempfile::TempDir;
-    use tokio::time::sleep;
-
-    #[tokio::test]
-    #[ignore = "abhi ke liye"]
-    async fn test_file_upload_single_binary_file() {
-        let temp_dir = TempDir::new().unwrap();
-        let upload_mw = file_upload(Some(FileUploadConfiguration {
-            upload_dir: temp_dir.path().to_string_lossy().to_string(),
-            ..Default::default()
-        }));
-
-        let mut req = HttpRequest::new();
-        let test_content = b"Hello, this is a test file content!";
-
-        // Set binary content using test method
-        req.set_binary(test_content.to_vec());
-        req.set_header("content-type", "application/octet-stream");
-
-        let res = HttpResponse::new();
-        let (req, _) = upload_mw(req, res).await;
-
-        // Check that file was saved
-        let uploaded_file = req.get_data("uploaded_file").unwrap();
-        let uploaded_path = req.get_data("uploaded_file_path").unwrap();
-
-        // Verify file exists and has correct content
-        assert!(
-            fs::metadata(&uploaded_path).is_ok(),
-            "File should exist at {}",
-            uploaded_path
-        );
-
-        // Read the file content and verify it matches
-        let file_content = fs::read_to_string(&uploaded_path).unwrap();
-        assert_eq!(
-            file_content, "Hello, this is a test file content!",
-            "File content mismatch. Expected: 'Hello, this is a test file content!', Got: '{}'",
-            file_content
-        );
-
-        // Check form data has the file reference
-        let form_data = req.form_data().unwrap();
-        assert_eq!(form_data.get("file"), Some(uploaded_file.as_str()));
-
-        // Check count
-        assert_eq!(req.get_data("uploaded_file_count"), Some("1".to_string()));
-
-        // Clean up
-        let _ = fs::remove_file(&uploaded_path);
-    }
-
-    #[tokio::test]
-    async fn test_file_upload_no_files() {
-        let temp_dir = TempDir::new().unwrap();
-        let upload_mw = file_upload(Some(FileUploadConfiguration {
-            upload_dir: temp_dir.path().to_string_lossy().to_string(),
-            ..Default::default()
-        }));
-
-        let mut req = HttpRequest::new();
-
-        // Create multipart form data with only text fields
-        let boundary = "----WebKitFormBoundary7MA4YWxkTrZu0gW";
-        let multipart_data = format!(
-            "--{boundary}\r\n\
-            Content-Disposition: form-data; name=\"name\"\r\n\
-            \r\n\
-            John Doe\r\n\
-            --{boundary}\r\n\
-            Content-Disposition: form-data; name=\"age\"\r\n\
-            \r\n\
-            30\r\n\
-            --{boundary}--\r\n"
-        );
-
-        req.set_binary(multipart_data.into_bytes());
-        req.set_header(
-            "content-type",
-            &format!("multipart/form-data; boundary={}", boundary),
-        );
-
-        let res = HttpResponse::new();
-        let (req, _) = upload_mw(req, res).await;
-
-        // Check text fields are accessible
-        let form_data = req.form_data().unwrap();
-        assert_eq!(form_data.get("name"), Some("John Doe"));
-        assert_eq!(form_data.get("age"), Some("30"));
-
-        // Check file data - the middleware might treat text-only multipart as a file upload
-        // So we check that it handles it gracefully rather than expecting no files
-        let file_count = req.get_data("uploaded_file_count");
-        if file_count.is_some() {
-            // If files were created, they should be accessible
-            assert!(req.get_data("uploaded_files").is_some());
-            assert!(req.get_data("uploaded_file").is_some());
-            assert!(req.get_data("uploaded_file_path").is_some());
-        } else {
-            // If no files were created, these should be None
-            assert_eq!(req.get_data("uploaded_files"), None);
-            assert_eq!(req.get_data("uploaded_file"), None);
-            assert_eq!(req.get_data("uploaded_file_path"), None);
-        }
-    }
-
-    #[tokio::test]
-    async fn test_file_upload_invalid_multipart() {
-        let temp_dir = TempDir::new().unwrap();
-        let upload_mw = file_upload(Some(FileUploadConfiguration {
-            upload_dir: temp_dir.path().to_string_lossy().to_string(),
-            ..Default::default()
-        }));
-
-        let mut req = HttpRequest::new();
-
-        // Create invalid multipart data
-        let invalid_data = "This is not valid multipart data";
-
-        req.set_binary(invalid_data.as_bytes().to_vec());
-        req.set_header("content-type", "multipart/form-data; boundary=invalid");
-
-        let res = HttpResponse::new();
-        let (req, _) = upload_mw(req, res).await;
-
-        // Should handle gracefully without crashing
-        // Note: Invalid multipart might fall back to binary upload, so we check for graceful handling
-        // rather than specific counts
-        assert!(
-            req.get_data("uploaded_file_count").is_some()
-                || req.get_data("uploaded_file_count").is_none()
-        );
-    }
-
-    #[tokio::test]
-    async fn test_file_upload_empty_multipart() {
-        let temp_dir = TempDir::new().unwrap();
-        let upload_mw = file_upload(Some(FileUploadConfiguration {
-            upload_dir: temp_dir.path().to_string_lossy().to_string(),
-            ..Default::default()
-        }));
-
-        let mut req = HttpRequest::new();
-
-        // Create empty multipart data
-        let boundary = "----WebKitFormBoundary7MA4YWxkTrZu0gW";
-        let empty_multipart = format!("--{boundary}--\r\n");
-
-        req.set_binary(empty_multipart.into_bytes());
-        req.set_header(
-            "content-type",
-            &format!("multipart/form-data; boundary={}", boundary),
-        );
-
-        let res = HttpResponse::new();
-        let (req, _) = upload_mw(req, res).await;
-
-        // Should handle gracefully
-        // Note: Empty multipart might fall back to binary upload, so we check for graceful handling
-        // rather than specific counts
-        assert!(
-            req.get_data("uploaded_file_count").is_some()
-                || req.get_data("uploaded_file_count").is_none()
-        );
-    }
-
-    #[tokio::test]
-    async fn test_multipart_with_files_no_middleware() {
-        // This test simulates what happens when a multipart form with files is uploaded
-        // WITHOUT the file upload middleware. The system should:
-        // 1. Detect it's a multipart form
-        // 2. Parse the multipart data
-        // 3. Extract text fields and make them accessible via form_data()
-        // 4. Ignore file parts (since no middleware is processing them)
-
-        // Create multipart form data with both text fields and a file
-        let boundary = "----WebKitFormBoundary7MA4YWxkTrZu0gW";
-        let multipart_data = format!(
-            "--{boundary}\r\n\
-            Content-Disposition: form-data; name=\"name\"\r\n\
-            \r\n\
-            John Doe\r\n\
-            --{boundary}\r\n\
-            Content-Disposition: form-data; name=\"age\"\r\n\
-            \r\n\
-            30\r\n\
-            --{boundary}\r\n\
-            Content-Disposition: form-data; name=\"file\"; filename=\"test.txt\"\r\n\
-            Content-Type: text/plain\r\n\
-            \r\n\
-            file content\r\n\
-            --{boundary}--\r\n"
-        );
-
-        // Simulate the request building process that happens in from_hyper_request
-        // First, determine the content type
-        let content_type = crate::req::determine_content_type_request(&format!(
-            "multipart/form-data; boundary={}",
-            boundary
-        ));
-        assert_eq!(
-            content_type,
-            crate::req::body::RequestBodyType::MultipartForm
-        );
-
-        // Parse the multipart data to extract fields and file parts
-        let (fields, file_parts) =
-            crate::helpers::parse_multipart_form(multipart_data.as_bytes(), boundary);
-
-        // Verify that both text fields and file parts were parsed correctly
-        assert_eq!(fields.len(), 2);
-        assert_eq!(file_parts.len(), 1);
-
-        // Check text fields
-        let name_field = fields.iter().find(|(k, _)| k == "name");
-        let age_field = fields.iter().find(|(k, _)| k == "age");
-        assert_eq!(name_field.map(|(_, v)| v), Some(&"John Doe".to_string()));
-        assert_eq!(age_field.map(|(_, v)| v), Some(&"30".to_string()));
-
-        // Check file part
-        let file_part = &file_parts[0];
-        assert_eq!(file_part.1.as_ref().unwrap(), "file"); // field name
-
-        // Now simulate what happens in the request building logic:
-        // Since there ARE file parts, the system should preserve raw bytes as BINARY
-        // This means form_data() won't work directly, but the text fields are still accessible
-        // through the parsed fields we extracted above
-
-        // Create a mock request to simulate the behavior
-        let mut req = HttpRequest::new();
-
-        // Simulate the parsed fields being inserted into form data
-        // Since set_form expects &'static str, we'll set them manually
-        req.set_form("name", "John Doe", crate::req::body::RequestBodyType::FORM);
-        req.set_form("age", "30", crate::req::body::RequestBodyType::FORM);
-
-        // Verify the form fields are accessible
-        let retrieved_form_data = req.form_data().unwrap();
-        assert_eq!(retrieved_form_data.get("name"), Some("John Doe"));
-        assert_eq!(retrieved_form_data.get("age"), Some("30"));
-
-        // The file field should NOT be accessible as form data
-        assert_eq!(retrieved_form_data.get("file"), None);
-
-        // This test demonstrates that:
-        // 1. Multipart forms with files are correctly identified
-        // 2. Text fields are extracted and made accessible
-        // 3. File parts are ignored when no middleware is present
-        // 4. The form_data() method works for text fields
-    }
-
-    #[tokio::test]
-    async fn test_multipart_with_files_request_building() {
-        // This test simulates the actual HTTP request building process
-        // to verify that our fix works end-to-end
-
-        // Create multipart form data with both text fields and a file
-        let boundary = "----WebKitFormBoundary7MA4YWxkTrZu0gW";
-        let multipart_data = format!(
-            "--{boundary}\r\n\
-            Content-Disposition: form-data; name=\"name\"\r\n\
-            \r\n\
-            John Doe\r\n\
-            --{boundary}\r\n\
-            Content-Disposition: form-data; name=\"age\"\r\n\
-            \r\n\
-            30\r\n\
-            --{boundary}\r\n\
-            Content-Disposition: form-data; name=\"file\"; filename=\"test.txt\"\r\n\
-            Content-Type: text/plain\r\n\
-            \r\n\
-            file content\r\n\
-            --{boundary}--\r\n"
-        );
-
-        // Simulate the request building process step by step
-
-        // Step 1: Determine content type
-        let content_type = crate::req::determine_content_type_request(&format!(
-            "multipart/form-data; boundary={}",
-            boundary
-        ));
-        assert_eq!(
-            content_type,
-            crate::req::body::RequestBodyType::MultipartForm
-        );
-
-        // Step 2: Parse multipart data
-        let (fields, file_parts) =
-            crate::helpers::parse_multipart_form(multipart_data.as_bytes(), boundary);
-
-        // Step 3: Verify parsing results
-        assert_eq!(fields.len(), 2);
-        assert_eq!(file_parts.len(), 1);
-
-        // Step 4: Simulate the request body creation logic
-        let mut form_data = crate::req::body::FormData::new();
-        for (key, value) in fields {
-            form_data.insert(key, value);
-        }
-
-        // Step 5: Create the request body using our new method
-        let request_body = if !file_parts.is_empty() {
-            // Has files: use our new method that preserves both binary data and form fields
-            crate::req::body::RequestBody::new_binary_with_form_fields(
-                multipart_data.into_bytes().into(),
-                form_data,
-            )
-        } else {
-            // No files: use regular form data
-            crate::req::body::RequestBody::new_form(form_data)
-        };
-
-        // Step 6: Verify the request body properties
-        assert_eq!(
-            request_body.content_type,
-            crate::req::body::RequestBodyType::BINARY
-        );
-
-        // Step 7: Verify that form fields are accessible from the binary content
-        if let crate::req::body::RequestBodyContent::BinaryWithFields(_, stored_form_data) =
-            &request_body.content
-        {
-            assert_eq!(stored_form_data.get("name"), Some("John Doe"));
-            assert_eq!(stored_form_data.get("age"), Some("30"));
-            assert_eq!(stored_form_data.get("file"), None); // File field should not be in form data
-        } else {
-            panic!("Expected BinaryWithFields variant");
-        }
-
-        // This test verifies that our new architecture correctly handles multipart forms with files
-        // by preserving both the binary data (for middleware processing) and the form fields
-        // (for direct access via form_data())
-    }
-
-    fn run_cors_middleware(
-        method: HttpMethods,
-        config: Option<CorsConfig>,
-    ) -> (HttpRequest, Option<HttpResponse>) {
-        let mut req = HttpRequest::new();
-        req.method = method;
-        let res = HttpResponse::new();
-        let mw = cors(config);
-        futures::executor::block_on(mw(req, res))
-    }
-
-    #[test]
-    fn test_cors_headers_default_config() {
-        // For non-OPTIONS requests, the middleware adds headers but returns None
-        // to continue to the next handler. We need a different approach to test this.
-        // Consider modifying the test helper to return the modified response object
-        // or test with OPTIONS which returns Some(response).
-        let (_, maybe_res) = run_cors_middleware(HttpMethods::OPTIONS, None);
-        assert!(maybe_res.is_some());
-        let res = maybe_res.unwrap();
-
-        assert_eq!(res.headers.get("Access-Control-Allow-Origin"), Some("*"));
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Methods"),
-            Some("GET, POST, PUT, DELETE, OPTIONS, HEAD")
-        );
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Headers"),
-            Some("Content-Type, Authorization")
-        );
-        assert_eq!(res.headers.get("Access-Control-Allow-Credentials"), None);
-    }
-
-    #[test]
-    fn test_cors_headers_custom_config_with_credentials() {
-        let config = CorsConfig {
-            allowed_origin: "https://example.com",
-            allowed_methods: "GET, POST",
-            allowed_headers: "X-Custom-Header",
-            allow_credentials: true,
-        };
-        // Test with OPTIONS to get the response with headers
-        let (_, maybe_res) = run_cors_middleware(HttpMethods::OPTIONS, Some(config.clone()));
-        assert!(maybe_res.is_some());
-        let res = maybe_res.unwrap();
-
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Origin"),
-            Some("https://example.com")
-        );
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Methods"),
-            Some("GET, POST")
-        );
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Headers"),
-            Some("X-Custom-Header")
-        );
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Credentials"),
-            Some("true")
-        );
-    }
-
-    #[test]
-    fn test_cors_options_preflight_returns_response() {
-        let (_, maybe_res) = run_cors_middleware(HttpMethods::OPTIONS, None);
-        assert!(maybe_res.is_some());
-        let res = maybe_res.unwrap();
-        assert_eq!(res.status_code.as_u16(), 200);
-        assert_eq!(res.headers.get("Access-Control-Allow-Origin"), Some("*"));
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Methods"),
-            Some("GET, POST, PUT, DELETE, OPTIONS, HEAD")
-        );
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Headers"),
-            Some("Content-Type, Authorization")
-        );
-    }
-
-    #[test]
-    fn test_cors_options_preflight_with_credentials() {
-        let config = CorsConfig {
-            allowed_origin: "https://foo.com",
-            allowed_methods: "OPTIONS",
-            allowed_headers: "X-Token",
-            allow_credentials: true,
-        };
-        let (_, maybe_res) = run_cors_middleware(HttpMethods::OPTIONS, Some(config.clone()));
-        assert!(maybe_res.is_some());
-        let res = maybe_res.unwrap();
-        assert_eq!(res.status_code.as_u16(), 200);
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Origin"),
-            Some("https://foo.com")
-        );
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Methods"),
-            Some("OPTIONS")
-        );
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Headers"),
-            Some("X-Token")
-        );
-        assert_eq!(
-            res.headers.get("Access-Control-Allow-Credentials"),
-            Some("true")
-        );
-    }
-
-    #[tokio::test]
-    async fn test_logger_default_config() {
-        let logger_mw = logger(None);
-        let mut req = HttpRequest::new();
-        req.path = "/test".to_string();
-        req.method = HttpMethods::POST;
-        let res = HttpResponse::new();
-
-        // Test that the middleware runs without panicking
-        // and returns the expected values
-        let (returned_req, maybe_res) = logger_mw(req.clone(), res.clone()).await;
-
-        assert_eq!(returned_req.path, "/test");
-        assert_eq!(returned_req.method, HttpMethods::POST);
-        assert!(maybe_res.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_logger_custom_config() {
-        let logger_mw = logger(Some(LoggerConfig {
-            method: true,
-            path: false,
-            ..Default::default()
-        }));
-
-        let mut req = HttpRequest::new();
-        req.path = "/foo".to_string();
-        req.method = HttpMethods::PUT;
-        let res = HttpResponse::new();
-
-        let (returned_req, maybe_res) = logger_mw(req.clone(), res.clone()).await;
-
-        assert_eq!(returned_req.path, "/foo");
-        assert_eq!(returned_req.method, HttpMethods::PUT);
-        assert!(maybe_res.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_logger_preserves_request_data() {
-        let logger_mw = logger(None);
-        let mut req = HttpRequest::new();
-        req.path = "/api/users".to_string();
-        req.method = HttpMethods::GET;
-        let res = HttpResponse::new();
-
-        let (returned_req, _) = logger_mw(req.clone(), res.clone()).await;
-
-        // Verify the middleware preserves all request data
-        assert_eq!(returned_req.path, req.path);
-        assert_eq!(returned_req.method, req.method);
-    }
-
-    #[tokio::test]
-    async fn test_logger_with_all_disabled() {
-        let logger_mw = logger(Some(LoggerConfig {
-            method: false,
-            path: false,
-            ..Default::default()
-        }));
-
-        let mut req = HttpRequest::new();
-        req.path = "/disabled".to_string();
-        req.method = HttpMethods::DELETE;
-        let res = HttpResponse::new();
-
-        let (returned_req, maybe_res) = logger_mw(req.clone(), res.clone()).await;
-
-        assert_eq!(returned_req.path, "/disabled");
-        assert_eq!(returned_req.method, HttpMethods::DELETE);
-        assert!(maybe_res.is_none());
-    }
-
-    fn mock_req(ip: &str) -> HttpRequest {
-        HttpRequest {
-            ip: IpAddr::from_str(ip).unwrap(),
-            headers: RequestHeaders::new(),
-            ..Default::default()
-        }
-    }
-
-    fn mock_req_with_header(ip: &str, header: (&str, &str)) -> HttpRequest {
-        let mut headers = HashMap::new();
-        headers.insert(header.0.to_string(), header.1.to_string());
-        HttpRequest {
-            ip: IpAddr::from_str(ip).unwrap(),
-            headers: RequestHeaders::_from_map(headers),
-            ..Default::default()
-        }
-    }
-
-    fn mock_res() -> HttpResponse {
-        HttpResponse::default()
-    }
-
-    #[tokio::test]
-    async fn allows_requests_within_limit() {
-        let mw = rate_limiter(Some(RateLimiterConfig {
-            max_requests: 3,
-            window_ms: Duration::from_millis(1000),
-            ..Default::default()
-        }));
-
-        let req = mock_req("127.0.0.1");
-        let res = mock_res();
-
-        // 1st request
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-
-        // 2nd request
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-
-        // 3rd request
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-    }
-
-    #[tokio::test]
-    async fn blocks_requests_over_limit() {
-        let mw = rate_limiter(Some(RateLimiterConfig {
-            max_requests: 2,
-            window_ms: Duration::from_millis(1000),
-            message: "Rate limit exceeded".to_string(),
-            ..Default::default()
-        }));
-
-        let req = mock_req("127.0.0.2");
-        let res = mock_res();
-
-        // 1st request
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-
-        // 2nd request
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-
-        // 3rd request should be blocked
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_some());
-        let resp = resp.unwrap();
-        assert_eq!(
-            resp.status_code,
-            crate::res::response_status::StatusCode::TooManyRequests
-        );
-        assert_eq!(
-            resp.headers.get("Retry-After").map(|v| v.to_string()),
-            Some("0".to_string())
-        );
-        assert_eq!(
-            resp.headers
-                .get("X-RateLimit-Remaining")
-                .map(|v| v.to_string()),
-            Some("0".to_string())
-        );
-    }
-
-    #[tokio::test]
-    async fn resets_after_window() {
-        let mw = rate_limiter(Some(RateLimiterConfig {
-            max_requests: 1,
-            window_ms: Duration::from_millis(100),
-            ..Default::default()
-        }));
-
-        let req = mock_req("127.0.0.3");
-        let res = mock_res();
-
-        // 1st request
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-
-        // 2nd request should be blocked
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_some());
-
-        // Wait for window to expire
-        sleep(Duration::from_millis(120)).await;
-
-        // 3rd request should be allowed again
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-    }
-
-    #[tokio::test]
-    async fn uses_proxy_header_when_enabled() {
-        let mw = rate_limiter(Some(RateLimiterConfig {
-            max_requests: 1,
-            window_ms: Duration::from_millis(1000),
-            proxy: true,
-            ..Default::default()
-        }));
-
-        let req = mock_req_with_header("127.0.0.4", ("X-Forwarded-For", "8.8.8.8"));
-        let res = mock_res();
-
-        // 1st request from 8.8.8.8
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-
-        // 2nd request from 8.8.8.8 should be blocked
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_some());
-    }
-
-    #[tokio::test]
-    async fn sets_rate_limit_headers() {
-        let mw = rate_limiter(Some(RateLimiterConfig {
-            max_requests: 2,
-            window_ms: Duration::from_millis(1000),
-            ..Default::default()
-        }));
-
-        let req = mock_req("127.0.0.5");
-        let res = mock_res();
-
-        // 1st request
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-
-        // 2nd request
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        assert!(resp.is_none());
-
-        // 3rd request should be blocked and have headers
-        let (_req, resp) = mw(req.clone(), res.clone()).await;
-        let resp = resp.unwrap();
-        assert_eq!(
-            resp.headers.get("X-RateLimit-Limit").map(|v| v.to_string()),
-            Some("2".to_string())
-        );
-        assert_eq!(
-            resp.headers
-                .get("X-RateLimit-Remaining")
-                .map(|v| v.to_string()),
-            Some("0".to_string())
-        );
-        assert!(resp.headers.get("X-RateLimit-Reset").is_some());
-        assert!(resp.headers.get("Retry-After").is_some());
-    }
-
-    #[test]
-    fn test_should_compress_content_type() {
-        assert!(should_compress_content_type("text/html"));
-        assert!(should_compress_content_type("text/html; charset=utf-8"));
-        assert!(should_compress_content_type("application/json"));
-        assert!(should_compress_content_type("TEXT/PLAIN")); // Case insensitive
-
-        assert!(!should_compress_content_type("image/jpeg"));
-        assert!(!should_compress_content_type("image/png"));
-        assert!(!should_compress_content_type("application/octet-stream"));
-        assert!(!should_compress_content_type("video/mp4"));
-    }
-
-    #[test]
-    fn test_compress_data() {
-        let original = b"Hello, World! ".repeat(100);
-        let compressed = compress_data(&original, 6).unwrap();
-
-        // Compressed data should be smaller than original for repetitive content
-        assert!(compressed.len() < original.len());
-
-        // Should have gzip magic numbers at the beginning
-        assert_eq!(&compressed[0..2], &[0x1f, 0x8b]);
-    }
-
-    #[test]
-    fn test_compression_config_default() {
-        let config = CompressionConfig::default();
-        assert_eq!(config.threshold, 1024);
-        assert_eq!(config.level, 6);
-    }
-
-    const DEFAULT_BODY_LIMIT: usize = 1024 * 1024; // 1 MB
-
-    fn make_req_with_body(body: Vec<u8>) -> HttpRequest {
-        HttpRequest {
-            body: RequestBody::new_binary(body),
-            ..Default::default()
-        }
-    }
-
-    fn make_res() -> HttpResponse {
-        HttpResponse::default()
-    }
-
-    #[tokio::test]
-    async fn test_body_within_limit() {
-        let limit = 10;
-        let body = vec![1, 2, 3, 4, 5];
-        let req = make_req_with_body(body.clone());
-        let res = make_res();
-
-        let middleware = body_limit(Some(limit));
-        let (req_out, resp_opt) = middleware(req, res).await;
-
-        assert_eq!(
-            req_out.body.content,
-            RequestBodyContent::BINARY(body.into())
-        );
-        assert!(resp_opt.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_body_exceeds_limit() {
-        let limit = 5;
-        let body = vec![1, 2, 3, 4, 5, 6, 7];
-        let req = make_req_with_body(body.clone());
-        let res = make_res();
-
-        let middleware = body_limit(Some(limit));
-        let (req_out, resp_opt) = middleware(req, res).await;
-
-        assert_eq!(
-            req_out.body.content,
-            RequestBodyContent::BINARY(body.into())
-        );
-        assert!(resp_opt.is_some());
-        let resp = resp_opt.unwrap();
-        assert_eq!(resp.status_code, StatusCode::PayloadTooLarge);
-    }
-
-    #[tokio::test]
-    async fn test_default_limit() {
-        let default_limit = DEFAULT_BODY_LIMIT;
-        let body = vec![0u8; default_limit + 1];
-        let req = make_req_with_body(body.clone());
-        let res = make_res();
-
-        let middleware = body_limit(None);
-        let (req_out, resp_opt) = middleware(req, res).await;
-
-        assert_eq!(req_out.body.content.len(), default_limit + 1);
-        assert!(resp_opt.is_some());
-        let resp = resp_opt.unwrap();
-        assert_eq!(resp.status_code, StatusCode::PayloadTooLarge);
-    }
-
-    #[tokio::test]
-    async fn test_body_exactly_at_limit() {
-        let limit = 8;
-        let body = vec![1u8; limit];
-        let req = make_req_with_body(body.clone());
-        let res = make_res();
-
-        let middleware = body_limit(Some(limit));
-        let (req_out, resp_opt) = middleware(req, res).await;
-
-        assert_eq!(
-            req_out.body.content,
-            RequestBodyContent::BINARY(body.into())
-        );
-        assert!(resp_opt.is_none());
-    }
-}
+pub mod cors;
+pub mod file_upload;
+pub mod logger;
+pub mod rate_limiter;
